[19] Reuse search result table  for query search and add helpers
