+++ assets/css/screens/_report.scss
@@ -1,14 +1,15 @@
-// TODO: Scope this to `body.report`, right now body class is `report-<id>`
-.report-summary {
-  &__container {
-    @apply py-8 px-4;
+body.report {
+  .report-summary {
+    &__container {
+      @apply py-8 px-4;
+    }
+
+    &__row {
+      @apply flex flex-wrap -mx-4;
+    }
   }
 
-  &__row {
-    @apply flex flex-wrap -mx-4;
+  .report-search-results {
+    @apply mt-10 mx-4;
   }
-}
-
-.report-search-results {
-  @apply mt-10 mx-4;
 }
