+++ lib/g_searcher_web/templates/report/show.html.eex
@@ -52,5 +52,5 @@
 </section>
 
 <section class="report-search-results">
-  <%= render "_search-result-table.html", search_results: @search_results, conn: @conn %>
+  <%= render GSearcherWeb.SharedView, "_search-result-table.html", assigns %>
 </section>
