+++ lib/g_searcher_web/templates/search_result/index.html.eex
@@ -1,6 +1,3 @@
-<%# TODO: Style this page in #19 %>
-<ul>
-  <%= for search_result <- @search_results do %>
-    <li><%= search_result.search_term %></li>
-  <% end %>
-</ul>
+<section class="search-result-list">
+  <%= render GSearcherWeb.SharedView, "_search-result-table.html", assigns %>
+</section>
