+++ lib/g_searcher_web/templates/shared/_search-result-table.html.eex
