+++ lib/g_searcher_web/views/layout_view.ex
@@ -1,10 +1,13 @@
 defmodule GSearcherWeb.LayoutView do
   use GSearcherWeb, :view
 
-  def class_name(conn) do
-    join_path(conn.path_info)
-  end
+  def body_class_name(conn),
+    do: "#{module_class_name(conn)} #{Phoenix.Controller.action_name(conn)}"
 
-  defp join_path(path) when is_list(path), do: Enum.join(path, "-")
-  defp join_path(path), do: path
+  defp module_class_name(conn) do
+    conn
+    |> Phoenix.Controller.controller_module()
+    |> Phoenix.Naming.resource_name("Controller")
+    |> String.replace("_", "-")
+  end
 end
