+++ test/g_searcher_web/views/layout_view_test.exs
@@ -1,8 +1,16 @@
 defmodule GSearcherWeb.LayoutViewTest do
   use GSearcherWeb.ConnCase, async: true
 
-  # When testing helpers, you may want to import Phoenix.HTML and
-  # use functions such as safe_to_string() to convert the helper
-  # result into an HTML string.
-  # import Phoenix.HTML
+  alias GSearcherWeb.LayoutView
+
+  describe "body_class_name/1" do
+    test "returns CSS class for page body with controller name", %{conn: conn} do
+      conn =
+        conn
+        |> put_private(:phoenix_controller, HelloWorld)
+        |> put_private(:phoenix_action, :show)
+
+      assert LayoutView.body_class_name(conn) == "hello-world show"
+    end
+  end
 end
